requirements.txt
streamlit>=1.36.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
nibabel>=5.2.0
opencv-python-headless>=4.9.0.80
matplotlib>=3.8.0
Pillow>=10.0.0
reportlab>=4.2.0
